add created/removed bones result reporting
