+++ io_scene_xray/ops/__init__.py
@@ -1,0 +1,9 @@
+import bpy
+
+class BaseOperator(bpy.types.Operator):
+    report_catcher = None
+
+    def __getattribute__(self, item):
+        if (item == 'report') and (self.report_catcher is not None):
+            return self.report_catcher
+        return super().__getattribute__(item)
